Yet another solving algorithm

+++ gr_object.h
@@ -17,7 +17,7 @@ struct GrShape {
         ClosePath,
         Text
     } type;
-    uint32_t x1,y1;
+    double x1,y1;
     //TODO: удалить поле
     uint32_t x2,y2;//ненужное т.к. все точки есть в векторе GrShape-ов
     uint32_t options;
